requirements.txt
Django>=2.2.3,<2.3.0
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.8.4,<2.9.0
django-filter>=2.2.0,<2.3.0

flake8>=3.7.8,<3.8.0